Add ExecutableJarPackageType
This is a type that copies a jar file from the artifact onto the host
and executes it, with provided arguments. This can be used to e.g. run
database migrations. It then cleans up the temporary directory it
created.

This package type requires a couple of new tasks, namely:

RmTmpdir - deletes a directory (only deletes from within /tmp/ as a
security measuer).

ExecuteJar - uses java -jar to execute the provided jar file.